# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SIS Kernel Control Daemon (sisctl)",
    "description": "REST API for managing SIS kernel QEMU instances",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8871",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/v1/graph/add-channel": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Add a channel to a graph",
        "operationId": "graphAddChannel",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddChannelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Channel added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddChannelResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/add-operator": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Add an operator to a graph",
        "operationId": "graphAddOperator",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddOperatorRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operator added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddOperatorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/start": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Start a graph execution",
        "operationId": "graphStart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGraphRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Graph started",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartGraphResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/predict": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Predict operator performance",
        "operationId": "graphPredict",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Prediction result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/feedback": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Submit feedback for operator",
        "operationId": "graphFeedback",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Feedback recorded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/state": {
      "get": {
        "tags": [
          "graph"
        ],
        "summary": "Get graph state",
        "operationId": "graphState",
        "parameters": [
          {
            "name": "graphId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Graph state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphState"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/graph/export": {
      "post": {
        "tags": [
          "graph"
        ],
        "summary": "Export graph as JSON",
        "operationId": "graphExport",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportGraphRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Graph exported",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportGraphResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddChannelRequest": {
        "type": "object",
        "properties": {
          "graphId": {
            "type": "string"
          },
          "cap": {
            "type": "integer"
          }
        },
        "required": [
          "graphId",
          "cap"
        ]
      },
      "AddChannelResponse": {
        "type": "object",
        "properties": {
          "channelId": {
            "type": "string"
          }
        },
        "required": [
          "channelId"
        ]
      },
      "AddOperatorRequest": {
        "type": "object",
        "properties": {
          "graphId": {
            "type": "string"
          },
          "opId": {
            "type": "string"
          },
          "in": {
            "type": "string"
          },
          "out": {
            "type": "string"
          },
          "prio": {
            "type": "integer"
          },
          "stage": {
            "type": "string"
          },
          "inSchema": {
            "type": "string"
          },
          "outSchema": {
            "type": "string"
          }
        },
        "required": [
          "graphId",
          "opId"
        ]
      },
      "AddOperatorResponse": {
        "type": "object",
        "properties": {
          "operatorId": {
            "type": "string"
          }
        },
        "required": [
          "operatorId"
        ]
      },
      "StartGraphRequest": {
        "type": "object",
        "properties": {
          "graphId": {
            "type": "string"
          },
          "steps": {
            "type": "integer"
          }
        },
        "required": [
          "graphId",
          "steps"
        ]
      },
      "StartGraphResponse": {
        "type": "object",
        "properties": {
          "started": {
            "type": "boolean"
          }
        },
        "required": [
          "started"
        ]
      },
      "PredictRequest": {
        "type": "object",
        "properties": {
          "opId": {
            "type": "string"
          },
          "latency_us": {
            "type": "integer"
          },
          "depth": {
            "type": "integer"
          },
          "prio": {
            "type": "integer"
          }
        },
        "required": [
          "opId",
          "latency_us",
          "depth"
        ]
      },
      "PredictResponse": {
        "type": "object",
        "properties": {
          "predicted": {
            "type": "number"
          },
          "conf": {
            "type": "number"
          }
        }
      },
      "FeedbackRequest": {
        "type": "object",
        "properties": {
          "opId": {
            "type": "string"
          },
          "verdict": {
            "type": "string"
          }
        },
        "required": [
          "opId",
          "verdict"
        ]
      },
      "FeedbackResponse": {
        "type": "object",
        "properties": {
          "recorded": {
            "type": "boolean"
          }
        },
        "required": [
          "recorded"
        ]
      },
      "GraphOperator": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "prio": {
            "type": "integer"
          },
          "state": {
            "type": "string"
          }
        }
      },
      "GraphChannel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "cap": {
            "type": "integer"
          },
          "depth": {
            "type": "integer"
          }
        }
      },
      "GraphStats": {
        "type": "object",
        "properties": {
          "operator_count": {
            "type": "integer"
          },
          "channel_count": {
            "type": "integer"
          },
          "total_executions": {
            "type": "integer"
          }
        }
      },
      "GraphState": {
        "type": "object",
        "properties": {
          "operators": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphOperator"
            }
          },
          "channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GraphChannel"
            }
          },
          "stats": {
            "$ref": "#/components/schemas/GraphStats"
          }
        }
      },
      "ExportGraphRequest": {
        "type": "object",
        "properties": {
          "graphId": {
            "type": "string"
          },
          "format": {
            "type": "string"
          }
        },
        "required": [
          "graphId",
          "format"
        ]
      },
      "ExportGraphResponse": {
        "type": "object",
        "properties": {
          "json": {
            "type": "string"
          }
        },
        "required": [
          "json"
        ]
      },
...
    }
  },
...
}